edit explore page
