--- src/components/lab.js
@@ -1,16 +1,0 @@
-// <div id="messages-45" className="message-2qnXI6 cozyMessage-3V1Y8y groupStart-23k01U wrapper-2a6GCs cozy-3raOZG zalgo-jN1Ica"
-//      role="group" tabIndex="-1" aria-setsize="50" aria-posinset="46">
-//     <div className="contents-2mQqc9" role="document">
-//         <img src="https://cdn.discordapp.com/avatars/429508180503166979/28ef094115c8b1c331ade07d64aad434.png?size=128"
-//         aria-hidden="true" className="avatar-1BDn8e clickable-1bVtEA" alt=" ">
-//             <h2 className="header-23xsNx">
-//                 <span className="username-1A8OIy clickable-1bVtEA focusable-1YV_-H" aria-controls="popout_195" aria-expanded="false" role="button" tabIndex="0" style="color: rgb(226, 192, 170);">ProfessorBetty | ProjectSunboot</span>
-//                 <span className="timestamp-3ZCmNB">
-//                     <span aria-label="Today at 12:24 PM">Today at 12:24 PM</span>
-//                 </span>
-//             </h2>
-//         <div className="markup-2BOw-j messageContent-2qWWxC">
-//             I love that series i think the episode is called the pigman. Then the thumb nail has a pigman in it
-//         </div>
-//     </div>
-// </div>
